+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ## 1.13.0 (IN PROGRESS)
 * Added iconAlignment "baseline" for <AppIcon>'s in the results list (UIIN-719)
-
 * Increase limit for holdings to 1000. Refs UIIN-723.
+* Add query param to new request url. Part of UIIN-773.
 
 ## [1.12.1](https://github.com/folio-org/ui-inventory/tree/v1.12.1) (2019-09-26)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.12.0...v1.12.1)
 
 * Correctly import `<Field>` in `StatisticalCodeSettings`. Refs UIIN-725
+* Create Inventory settings page for HRID handling. Refs UIIN-741
 
 ## [1.12.0](https://github.com/folio-org/ui-inventory/tree/v1.12.0) (2019-09-12)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.11.1...v1.12.0)